+++ test/deprecated/test_parse_project_page.py
@@ -31,6 +31,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.1.0.tar.gz",
@@ -42,6 +43,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.1.0.post1-py3-none-any.whl",
@@ -53,6 +55,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.1.0.post1.tar.gz",
@@ -64,6 +67,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.2.0-py3-none-any.whl",
@@ -75,6 +79,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.2.0.tar.gz",
@@ -86,6 +91,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.3.0-py3-none-any.whl",
@@ -97,6 +103,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.3.0.tar.gz",
@@ -108,6 +115,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.4.0-py3-none-any.whl",
@@ -119,6 +127,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.4.0.tar.gz",
@@ -130,6 +139,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.4.1-py3-none-any.whl",
@@ -141,6 +151,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.4.1.tar.gz",
@@ -152,6 +163,7 @@ def test_parse_qypi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
     ]
 
@@ -175,6 +187,7 @@ def test_parse_qypi_base():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.1.0.tar.gz",
@@ -186,6 +199,7 @@ def test_parse_qypi_base():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.1.0.post1-py3-none-any.whl",
@@ -197,6 +211,7 @@ def test_parse_qypi_base():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.1.0.post1.tar.gz",
@@ -208,6 +223,7 @@ def test_parse_qypi_base():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
     ]
 
@@ -231,6 +247,7 @@ def test_parse_qypi_mixed():
             has_sig=None,
             yanked="Metadata was smelly",
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.1.0.tar.gz",
@@ -242,6 +259,7 @@ def test_parse_qypi_mixed():
             has_sig=None,
             yanked="",
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.1.0.post1-py3-none-any.whl",
@@ -253,6 +271,7 @@ def test_parse_qypi_mixed():
             has_sig=True,
             yanked="",
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.1.0.post1.tar.gz",
@@ -264,6 +283,7 @@ def test_parse_qypi_mixed():
             has_sig=True,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="qypi-0.2.0-py3-none-any.whl",
@@ -275,6 +295,7 @@ def test_parse_qypi_mixed():
             has_sig=False,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
     ]
 
@@ -298,6 +319,7 @@ def test_parse_devpi_devpi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="devpi-2.1.0.tar.gz",
@@ -309,6 +331,7 @@ def test_parse_devpi_devpi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="devpi-2.0.3.tar.gz",
@@ -320,6 +343,7 @@ def test_parse_devpi_devpi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="devpi-2.0.2.tar.gz",
@@ -331,6 +355,7 @@ def test_parse_devpi_devpi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
         DistributionPackage(
             filename="devpi-2.0.1.tar.gz",
@@ -342,5 +367,6 @@ def test_parse_devpi_devpi():
             has_sig=None,
             yanked=None,
             metadata_digests=None,
+            has_metadata=False,
         ),
     ]